fix_errors
