+++ app/(site)/layout.tsx
@@ -1,6 +1,6 @@
 import type {Metadata} from "next";
 import {Noto_Sans, Open_Sans} from "next/font/google";
-import "./globals.css";
+import "../globals.css";
 
 const notoSans = Noto_Sans({
 	variable: "--font-noto-sans",
